nafsia/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
transformers
torch
spacy
numpy
httpx
pydantic
websockets
